refactor: Cars 클래스 삭제 - RacingGameRunner에 반영
