+++ src/test/java/racingcar/MoveCarTest.java
@@ -4,9 +4,6 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 
-import java.util.Arrays;
-import java.util.List;
-
 public class MoveCarTest {
 
     @Test
@@ -19,10 +16,10 @@ public class MoveCarTest {
 
     @RepeatedTest(10)
     public void multiCarsMove(){
-        List<String> carNames = Arrays.asList("Car1", "Car2", "Car3");
-        Cars cars = new Cars(carNames);
-        cars.move();
-        for (Car car: cars.getCarList()){
+        String carNames = "Car1,Car2,Car3";
+        RacingGameRunner racingGameRunner = new RacingGameRunner(carNames,1);
+        racingGameRunner.move();
+        for (Car car: racingGameRunner.getCarList()){
             Assertions.assertThat(car.getPosition()).isBetween(0, 1);
         }
     }
